feat(develop): Death pop-up working (with respawn button)
Also, some sounds have been added to the game, the death animation of the zombie enemy has been created and some bugs have been solved

+++ Assets/Scripts/Punch.cs
@@ -8,7 +8,7 @@ public class Punch : MonoBehaviour
     bool isLaunching;
 
     [Header("Damage Attack")]
-    public float damageAttack = 2.0f; //half heart
+    public int damageAttack = 2; //half heart
 
     [Header("Particles")]
     public ParticleSystem punchHitEffect;
